+++ libraries/react-ui/src/components/PushToTalkButton.tsx
@@ -28,56 +28,78 @@ export type PushToTalkButtonProps = {
    * Optional "bottom" string turns on internal placement without any CSS positioning.
    */
   placement?: string
+
   /**
    * Keyboard key to use for controlling the button.
    * Passing e.g. ` ` (a spacebar) will mean that holding down the spacebar key will key the button pressed.
    */
   captureKey?: string
+
   /**
    * The size of the button, as CSS (e.g. `5rem`).
    */
   size?: string
+
   /**
    * Colours of the gradient around the button.
    * Valid input is an array of two hex colour codes, e.g. `['#fff', '#000']`.
    */
   gradientStops?: string[]
+
   /**
    * Optional boolean. Default: false
    */
   hide?: boolean
+
   /**
    * Optional string containing a short usage introduction. Displayed when the component is first displayed. Default: "Push to talk". Set to "" to disable.
    */
   intro?: string
+
   /**
    * Optional string containing a short usage hint. Displayed on a short tap. Default: "Push to talk". Set to "" to disable.
    */
   hint?: string
+
   /**
    * Optional CSS string for hint text. Default: "1.2rem"
    */
   fontSize?: string
+
   /**
    * Optional number in ms. Visibility duration for intro and hint callouts. Default: "5000" (ms)
    */
   showTime?: number
+
   /**
    * Optional string (CSS color) for hint text. Default: "#ffffff"
    */
   textColor?: string
+
   /**
    * Optional string (CSS color) for hint text background. Default: "#202020"
    */
   backgroundColor?: string
+
   /**
    * Optional boolean. Shows poweron state. If false, recording can immediately start but will first press will cause a system permission prompt. Default: false
    */
   powerOn?: boolean
+
   /**
    * Optional CSS string. Vertical distance from viewport edge. Only effective when using placement.
    */
   voffset?: string
+
+  /**
+   * Optional time in milliseconds to listen after tap. Set to 0 to disable tap-to-talk. Default: "8000" (ms)
+   */
+  tapToTalkTime?: number
+
+  /**
+   * Optional milliseconds of silence to listen before hangup. Only used in tap-to-talk mode. Default: "1000" (ms)
+   */
+  silenceToHangupTime?: number
 
 }
 
@@ -88,6 +110,11 @@ export type PushToTalkButtonProps = {
  *
  * @public
  */
+
+type IButtonState = {
+  holdListening: boolean
+  stopContextTimeout: number | null
+}
 
 export const PushToTalkButton: React.FC<PushToTalkButtonProps> = ({
   powerOn = false,
@@ -103,15 +130,21 @@ export const PushToTalkButton: React.FC<PushToTalkButtonProps> = ({
   backgroundColor,
   placement,
   voffset,
+  tapToTalkTime = 8000,
+  silenceToHangupTime = 1000,
 }) => {
-  const { speechState, toggleRecording, initialise } = useSpeechContext()
+  const { speechState, toggleRecording, initialise, segment } = useSpeechContext()
   const [icon, setIcon] = useState<string>((powerOn ? SpeechState.Idle : SpeechState.Ready) as string)
   const [hintText, setHintText] = useState<string>(intro)
   const [showHint, setShowHint] = useState(true)
+  const buttonStateRef = useRef<IButtonState>({
+    holdListening: false,
+    stopContextTimeout: null,
+  })
   const buttonRef = useRef<any>()
   const speechStateRef = useRef<SpeechState>()
 
-  const SHORT_PRESS_TRESHOLD_MS = 600
+  const TAP_TRESHOLD_MS = 600
 
   // make stateRef always have the current count
   // your "fixed" callbacks can refer to this object whenever
@@ -134,11 +167,12 @@ export const PushToTalkButton: React.FC<PushToTalkButtonProps> = ({
     if (!powerOn && speechState === SpeechState.Idle) {
       setIcon(SpeechState.Ready as string)
     } else {
+      console.log(speechState as string)
       setIcon(speechState as string)
     }
 
     // Automatically start recording if button held
-    if (!powerOn && buttonRef?.current?.isbuttonpressed() === true && speechState === SpeechState.Ready) {
+    if (!powerOn && (buttonRef?.current?.isbuttonpressed() === true || buttonStateRef.current.holdListening) && speechState === SpeechState.Ready) {
       toggleRecording().catch(err => console.error('Error while starting to record', err))
     }
   // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -148,6 +182,11 @@ export const PushToTalkButton: React.FC<PushToTalkButtonProps> = ({
     PubSub.publish(SpeechlyUiEvents.TangentPress, { state: speechStateRef.current })
     window.postMessage({ type: 'holdstart', state: mapSpeechStateToClientState(speechStateRef.current !== undefined ? speechStateRef.current : SpeechState.Idle) }, '*')
     setShowHint(false)
+
+    if (buttonStateRef.current.stopContextTimeout) {
+      window.clearTimeout(buttonStateRef.current.stopContextTimeout)
+      buttonStateRef.current.stopContextTimeout = null
+    }
 
     switch (speechStateRef.current) {
       case SpeechState.Idle:
@@ -168,25 +207,55 @@ export const PushToTalkButton: React.FC<PushToTalkButtonProps> = ({
     PubSub.publish(SpeechlyUiEvents.TangentRelease, { state: speechStateRef.current, timeMs: event.timeMs })
     window.postMessage({ type: 'holdend' }, '*')
 
-    switch (speechStateRef?.current) {
-      case SpeechState.Ready:
-      case SpeechState.Recording:
-      case SpeechState.Connecting:
-      case SpeechState.Loading:
-        if (event.timeMs < SHORT_PRESS_TRESHOLD_MS) {
-          setHintText(hint)
-          setShowHint(true)
+    if (event.timeMs < TAP_TRESHOLD_MS) {
+      if (tapToTalkTime === 0) {
+        setHintText(hint)
+        setShowHint(true)
+      } else {
+        // Short press when not recording = schedule "silence based stop"
+        if (!buttonStateRef.current.holdListening) {
+          setStopContextTimeout(tapToTalkTime)
         }
-        break
+      }
     }
 
-    switch (speechStateRef.current) {
-      case SpeechState.Recording:
-        toggleRecording().catch(err => console.error('Error while stopping recording', err))
-        break
-      default:
-        break
+    if (!buttonStateRef.current.stopContextTimeout) {
+      stopListening()
     }
+  }
+
+  const setStopContextTimeout = (timeoutMs: number): void => {
+    buttonStateRef.current.holdListening = true
+    if (buttonStateRef.current.stopContextTimeout) {
+      window.clearTimeout(buttonStateRef.current.stopContextTimeout)
+    }
+    buttonStateRef.current.stopContextTimeout = window.setTimeout(() => {
+      buttonStateRef.current.stopContextTimeout = null
+      stopListening()
+    }, timeoutMs)
+  }
+
+  const stopListening = (): void => {
+    buttonStateRef.current.holdListening = false
+    if (isStoppable(speechStateRef.current)) {
+      toggleRecording().catch(err => console.error('Error while stopping recording', err))
+    }
+  }
+
+  /**
+   * Extend listening time if segment updates received
+   */
+  useEffect(() => {
+    if (segment) {
+      if (buttonStateRef.current.stopContextTimeout) {
+        setStopContextTimeout(silenceToHangupTime)
+      }
+    }
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [segment])
+
+  const isStoppable = (s?: SpeechState): boolean => {
+    return (s === SpeechState.Recording)
   }
 
   return (